add UpdatePartClassDepend to readobject.c
