+++ Source/smokeview/smokeheaders.h
@@ -732,7 +732,6 @@ EXTERNCPP void Cam2Clip(cameradata *cam);
 EXTERNCPP void GetElevAz(float *xyznorm,float *dtheta, float *rotate_axis, float *dpsi);
 EXTERNCPP char *GetDeviceLabel(char *buffer);
 EXTERNCPP void DrawDevices(int mode);
-EXTERNCPP void UpdatePartClassDepend(partclassdata *partclassi);
 
 EXTERNCPP void CheckLab(void);
